fix: work around markdown java syntax issue
The latest vim and neovim versions have a bug which creates a circular
dependency between the java and markdown syntax files. This works around
the issue by making sure java highlighting is never invoked from a
markdown file. Removing from `g:markdown_fenced_languages` would achieve
a similar effect, but it's always possible that a later script or plugin
would append to this list and reintroduce the bug, so setting
`g:java_ignore_markdown` seems to be more robust.

It's possible to have a more complex workaround (documented in the
linked GitHub PR on the vim repo), however I can live without this
highlighting feature for now.